refactor: domain,view 폴더 분리
